WIP Issue #132 - Add some tests for the utils modules
Addressed initial feedback, removed unnecessary flags and empty tests.
