+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 
 # Artifact of building project
 build/
+
+.nyc_output
