--- .nyc_output/666ef0fbd5f3f9dacb403f64ffb3dad0.json
@@ -1,1 +1,0 @@
-{}
