--- .nyc_output/f92c904f750d23c1bdbad7044f72cda9.json
@@ -1,1 +1,0 @@
-{"C:\\Users\\Kathy Ren\\Desktop\\firefox-code-coverage-frontend\\.neutrinorc.js":{"path":"C:\\Users\\Kathy Ren\\Desktop\\firefox-code-coverage-frontend\\.neutrinorc.js","statementMap":{"0":{"start":{"line":1,"column":0},"end":{"line":33,"column":2}},"1":{"start":{"line":28,"column":6},"end":{"line":30,"column":9}},"2":{"start":{"line":29,"column":8},"end":{"line":29,"column":37}}},"fnMap":{"0":{"name":"(anonymous_0)","decl":{"start":{"line":27,"column":4},"end":{"line":27,"column":5}},"loc":{"start":{"line":27,"column":18},"end":{"line":31,"column":5}},"line":27},"1":{"name":"(anonymous_1)","decl":{"start":{"line":28,"column":66},"end":{"line":28,"column":67}},"loc":{"start":{"line":28,"column":76},"end":{"line":30,"column":7}},"line":28}},"branchMap":{},"s":{"0":1,"1":1,"2":0},"f":{"0":1,"1":0},"b":{},"_coverageSchema":"332fd63041d2c1bcb487cc26dd0d5f7d97098a6c","hash":"f0c1f8c33a4a3148a779d18b64cf3e74d46034b4","contentHash":"616afa8e6289ceed125fa6e287560c09_11.6.0"}}
