+++ test/mocha.opts
@@ -1,2 +1,1 @@
---require babel-core/register
 --recursive
